Interactive chart: Progressive enhancement at client side
Progressively enhance the chart rendering by using the ECharts library at the client side.
The chart is rendered as an image on the server side,
and the image is replaced with an interactive chart on the client side.

In addition to parser function, define a parser tag
and implement a webcomponent for that.

Features:
* Adds scrolling for legend
* Localization for numbers e.g. Hebrew
* Responsive design
* Adds interactive tooltip for data exploration

This is a follow up from various discussions at Indianapolis product and tech offsite.
This has since been discussed with Charts designer and PM.

Depends-On: https://gitlab.wikimedia.org/repos/mediawiki/services/chart-renderer/-/merge_requests/41
Bug: T378210
Change-Id: I5302fd4e534bda4f2d91bb791e683e64744e4c36