Add tracking categories for {{#chart:}} usage
Pages are added to a tracking category controllable via the message
"chart-render-category" and "chart-error-category" depending on
success. English names are "Pages using the Chart extension" and
"Pages using the Chart extension with rendering errors".

Error rendering is moved to use StatusFormatter from core.
There may be an easier way to get a LocalizationContext for rendering
content messages than is used here.

Bug: T369684
Change-Id: I9cba3672ff43cea7d8dbe4c5d3d6fe4fa2abf379