Support new data-chart attribute
https://gitlab.wikimedia.org/repos/mediawiki/services/chart-renderer/-/merge_requests/60
replaces the data-spec, data-theme, data-x-axis-type and
data-y-axis-type attributes with a single data-chart attribute, which is
encoded differently. Add code interpreting this new attribute.

Also keep support for the old attributes for backward compatibility,
so that pages with cached HTML won't break. Eventually, when the new
attribute has been deployed long enough for all pages with the old
attributes to have fallen out of the cache, we can get rid of the
backward compatibility code.

Add an ADR documenting the reasons for this change.

Also do other minor cleanup in the rendering code:
- The xFormatter and yFormatter params to specWithTooltip were marked as
  optional and could be passed in as undefined, but this would have
  errored. Make them required and ensure they're always passed in.
- Rename specWithTooltip to addTooltip and have it add to the spec
  object rather than creating a new object.
- Get rid of the wrapper function around renderInNode, and rename
  renderInNode to render.

Bug: T383613
Change-Id: Ia1a37d7cb570c22f5d7f7d70aaac1a89ad19de1c