Restore globaljsonlinks updates that got removed by mistake
Earlier refactor adding tracking categories accidentally removed
the recordJsonLinks calls which actually added the tracking info
to the parser output, breaking usage tracking on current installs.

This fix is needed to restore these calls during rendering and
repopulate the data.

Fixed as part of work on usage tracking display surfacing, which
confirmed this was broken earlier this week. :D

Bug: T383446
Change-Id: Ib5f0854e707cf6012801b3551c08ce5e1c7cdbbb