ADR for Scribunto Lua transforms on Charts+JsonConfig
Indicates that we're adding Scribunto-based Lua transforms, why, and
where (in JsonConfig for base plus here for the usage)

Added some detail on proposed arguments passing (a JSON object to be
converted to Lua and passed in, usually expected to be a map of named
arguments or empty)

Clarified what happens if specified functions break, and the mechanics
of the underlying transfer of information from the store on Commons to
the other wikis as consumers of centrally-provided data+transforms.

Note that we will want to expose the dependency backlinks in some way
to make it easier to track down potential protections or vandalism
hits that may affect a page, but we've got them defined as going into
globaljsonlinks so we'll have the records here. :)

JsonConfig side:

* JCTransform/JCTransformOutput will encapsulate the logic
* JCSingleton::getContent() is expanded with transform option
* action=jsondata in API is expanded with transform option and props
* globaljsonlinks will store dependencies for Module: pages as well
  as Data: pages

Charts side:

* Chart format syntax is expanded to specify transform and args
* Arguments are currently not overrideable at invocation
* Charts uses the expanded JCSingleton::getContent()
* Example using a logic module and data set shared between two
  temperature charts using different units
* Charts will record the returned dependencies to JsonConfig's
  globaljsonlinks on page render

Bug: T388579
Change-Id: I9caaa11a1ad4805f3b5aca913e38b3af2162c6d7