Initial fix for mismatched user language on Chart
We were using the user language for client-side formatting, but with
data that was localized for the global content or page-specific target
language. This could cause visible mismatches in numeric formatting
etc.

Asking config for wgPageViewLanguage appears to give us the per-page
setting correctly, so the formatting matches with the data and label
localizations.

Doesn't make any rendering changes yet.

To render in another language on a multilingual site like Commons, set
the page language for the target page using the various existing
methods for doing that and they should come up correctly now both for
the localized texts and the formatting.

Note that this means uselang=XX will not affect a chart's appearance,
as it's tied to page or content language.

Bug: T383482
Change-Id: I271957492def6af65fdb9e9e352fc2802e090725