Fix localization for validation errors checking tabular data
If the referenced tabular data source on a .chart page is missing
or invalid, we show an error message from the .chart page's validation
*and* we show an error message indicating the the chart page was
invalid in general.

Previously the validation details were left out of the localization
arrays, giving an erroneous display of a broken message. This is fixed
here by adding the missing messages from ChartSourceValidator.

Bug: T389126
Change-Id: I719c9fc03f337a3b8f228a01ad1cb122b767fb54