Specify Lua transform arguments on {{#chart:}} invocations
Allows overriding transform arguments by specifying additional
key=value pairs in the `{{#chart:}}` invocation directly, with
an "arg:" prefix.

These will be applied on top of whatever is specified in the
`transform.args` map in the chart format JSON.

Added some test cases for ChartArgumentsParser exercising this,
and initial integration tests for the parserfunction which simply
confirm that errors are handled gracefully without fatal/throw.

Bug: T395610
Change-Id: Iba31c87ecba7608a6efd39aa74e542eac931fc22