Fix for variant language handling in Charts
The base target language we were seeing was the base language without
a variant (whoops), had to dive a little deeper to get the target
language _with_ variant _during_ the parse.

It's available from ParserOutput->getLanguage() _after_ the parse
but we needed it earlier, so doing the same thing that Parser does
when setting up the ParserOutput and grabbing the Parser's own
getTargetLanguageConverter and asking it for its preferred
variant, as derived here from a simulated uselang param.

Test case runs parser with a stubbed-out renderer that returns
the language, seems to confirm if target language is set that
its value is now passed through.

Bug: T394366
Change-Id: Iac91bfb3bd50f9644a26da8e3ce643faa4b44eef