Robustness for cache on Charts in languages with variants
The target language's preferred variant is set by the global user
language but this wasn't being marked on the cache keys by calling
the proper accessor to indicate user language was relevant.

If the target language has variants, bifurcate cache on user language
to ensure that it is correctly marked up on the ParserOptions.

Bug: T396913
Change-Id: I070ac1308f64ae9fb0b1e81ec613d6b23e7e8110