Refactoring validation and localization
Reduces redundancy in setup of new fields in JCChartContent;
now instead of having to add to both validate and localize
methods to copy the same data around, a single common array
is set up in self::VALIDATION_SCHEMA which lists the various
fields and their validators, and whether they need to be
copied with localization or raw.

Extra code is only necessary for additional logic, such as
ensuring that the transform module and function are both
present if a transform is set, even though you can leave
out the whole 'transform' section safely.

Bug: T374061
Change-Id: Ib230cf0bc10fb358368284b154d3d3b5e00135f6