Re-expose server-side rendering if client-side render fails
Previously if we detected an error during echarts rendering we
would log it, and if a network error prevented loading the module
we wouldn't do anything, and in either case the fallback rendering
would remain blurred out.

Now, if a failure is caught from the using() or in the render()
it will be logged (if relevant) and the fallback render unblurred.

Bug: T382968
Change-Id: I19eb7e6d4696b54167957e21147ad30bc10b2d51