VisualEditor: Treat charts as block, not inline
Charts have always been block elements (the `<wiki-chart>` tag is styled
with `display: block`), but for some reason we were using
MWTransclusionInlineNode to recognize them. This is wrong, use
MWTransclusionBlockNode instead.

This also causes error boxes to be correctly recognized as charts and
become editable as such.

Also remove the check for data-mw-charts: this didn't work because it
was a typo (the correct attribute name is data-mw-chart), and it would
have exploded had it been spelled correctly because the JSON blobs in
data-mw and data-mw-chart are very different. Just use data-mw, which is
what we need in this case.

Bug: T393308
Change-Id: I31efd1cfc414e3fceaf64c5ccb8b9c1a4e28f446