Refactor isSwitchableString to correctly report details on invalid
Instead of just returning one or the other of the localized string
or string line error results on failure, return a combined description
that indicates either is possible and includes the max length.

Uses the underlying helper functions from JsonConfig instead of calling
through to the high-level validators, and adds a string.

Bug: T394959
Change-Id: I829183b8598d74ad819f999fcc04189f038c160e