API action=chartinfo internal helper for Charts stats
Pulls up the number of unique-per-page "Data:*.chart" usages via
Charts. Can show either local wiki or all wikis figure.

We expect to fetch data from this on the live wikis to record
usage over time. An hour-TTL cache is placed in front to
reduce brigading behavior if it gets expensive for now,
but consider adding a suffix column to globaljsonlinks_target
with an additional index if this becomes a problem.

Pass "global=1" to include counts for all wikis instead of
just the current.

Bug: T393950
Depends-On: I9eedb171a265fc959abe79c5b0a2cfbd0782dd7e
Change-Id: I760355b5b0da1d82bdf6da735ebe6c8125717705