Fix for validation error display in transformed chart data
This fixes the display of validation errors on charts that are run
through Lua ransforms. Mishandling of "good" vs "ok" Status objects
led to passing a null Content into rendering and breaking at a high
level in rendering.

For charts containing null values this is currently _all of them_ due
to a known limitation with transferring data between PHP and Lua.

This gets general page rendering fixed and displays the errors,
indicating the short rows from the Lua stuff. These still need to be
fixed in the transform handling or, possibly, in Scribunto, but now
they won't break rendering.

Bug: T398597
Change-Id: I99ec05e5373cd562569a4f0b5c0c1ad69caff83d